processors/sr/sr/requirements.txt
absl-py==2.1.0
affine==2.4.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
ftfy==6.2.3
GDAL==3.6.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.66.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pip==24.2
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
rasterio==1.3.11
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==75.1.0
six==1.16.0
snuggs==1.4.7
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.9
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
zipp==3.20.2
